+++ Network/API/Telegram/Bot/Object/Update/Message/Content/File.hs
@@ -5,13 +5,7 @@ import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Document as E
 import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Photo as Exports
 import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Video as Exports
 import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Voice as Exports
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Caption as Exports
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Duration as Exports
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Filesize as Exports
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.MIME as Exports
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Height as Exports
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Width as Exports
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.URI as Exports
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special as Exports
 
 import "aeson" Data.Aeson (FromJSON (parseJSON), withObject, (.:))
 import "base" Control.Applicative ((<|>))