+++ Network/API/Telegram/Bot/Object/Update/Message/Content/File/Special.hs
@@ -1,0 +1,9 @@
+module Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special (module Exports) where
+
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Caption as Exports
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Duration as Exports
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Filesize as Exports
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.MIME as Exports
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Height as Exports
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Width as Exports
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.URI as Exports